pyproject.toml
[project]
name = "openapi-gateway"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "docker>=7.0",
    "python-jose[cryptography]>=3.3",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gateway"]
